🔨 Refactor: queries 하위 디렉토리 이름 변경
